prefer apps based on their usage count
count how many times each app was selected and also record it in the
history, then sort the apps based on this number so that most used
app is the first suggested.

changes structure of the history file so instead of just app name
on each line it contains "<app name>:<count_int>", if there
is no ":<count_int>" it assumes old format and fallbacks
to count=1.